refactoring this

+++ src/main/java/models/Sale.java
@@ -38,7 +38,7 @@ public class Sale {
         this.productId = productId;
         int numPurchased = document.getLong(fields.numPurchased).intValue();
         this.numPurchased = numPurchased;
-        double cost = document.getLong(fields.cost).doubleValue();
+        double cost = document.getDouble(fields.cost).doubleValue();
         this.cost = cost;
         this.documentReference = getSaleDocument();
     }
